#16 - Refactor code, add GlobalExceptionHandler for readability of controllers

+++ src/main/java/org/sviatdev/lecturesharing/services/UserService.java
@@ -1,11 +1,16 @@
 package org.sviatdev.lecturesharing.services;
 
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.sviatdev.lecturesharing.dao.UserDao;
+import org.sviatdev.lecturesharing.exceptions.UserNotFoundException;
 import org.sviatdev.lecturesharing.models.University;
 import org.sviatdev.lecturesharing.models.User;
+
+import java.util.List;
+import java.util.Optional;
+
+import static org.springframework.http.HttpStatus.*;
 
 @Service
 public class UserService {
@@ -16,48 +21,51 @@ public class UserService {
         this.userDao = userDao;
     }
 
-    public ResponseEntity<?> getAllUsers() {
+    public List<User> getAllUsers() throws UserNotFoundException {
         var users = userDao.findAll();
-        return !users.isEmpty()
-                ? ResponseEntity.ok(users)
-                : ResponseEntity.status(HttpStatus.NOT_FOUND).body("No users found.");
-    }
-
-    public ResponseEntity<?> insertUser(User user) {
-        if (isUserExist(user)) {
-            return ResponseEntity.status(HttpStatus.CONFLICT).body("User already exists.");
+        if (users.isEmpty()) {
+            throw new UserNotFoundException();
         }
-        return ResponseEntity.ok(userDao.save(user));
+        return users;
     }
 
-    public ResponseEntity<?> getUserById(Long id) {
-        var user = userDao.findById(id);
-        return user.isPresent()
-                ? ResponseEntity.ok(user)
-                : ResponseEntity.status(HttpStatus.NOT_FOUND).body("No users found.");
+    public Optional<User> getUserById(Long id) {
+        return userDao.findById(id);
+    }
+
+    public List<User> findUsersByUniversity(University university) {
+        if (university == null) {
+            throw new IllegalArgumentException();
+        }
+        return userDao.findByUniversity(university);
+    }
+
+    public void insertUser(User user) throws UserNotFoundException {
+        if (user == null || user.getUsername() == null || user.getPassword() == null) {
+            throw new IllegalArgumentException();
+        } else if (!isUserExist(user)) {
+            throw new UserNotFoundException();
+        }
+        userDao.save(user);
     }
 
     public ResponseEntity<?> findByUsername(String userName) {
         var user = userDao.findByUsername(userName);
         return user.isPresent()
                 ? ResponseEntity.ok(user)
-                : ResponseEntity.status(HttpStatus.NOT_FOUND).body("No user found.");
+                : ResponseEntity.status(NOT_FOUND).body("No user found.");
     }
 
-    public ResponseEntity<?> findUsersByUniversity(University university) {
-        var users = userDao.findUsersByUniversity(university);
-        return !users.isEmpty()
-                ? ResponseEntity.ok(users)
-                : ResponseEntity.status(HttpStatus.NOT_FOUND).body("No users found.");
-    }
-
-    public ResponseEntity<?> removeUser(Long id) {
-        try {
-            userDao.deleteById(id);
-            return ResponseEntity.ok("User deleted successfully.");
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.INTERNAL_SERVER_ERROR).body("An error occurred.");
+    public void removeUser(Long id) throws UserNotFoundException {
+        if (id == null) {
+            throw new IllegalArgumentException();
         }
+
+        var user = getUserById(id);
+        if (user.isEmpty()) {
+            throw new UserNotFoundException();
+        }
+        userDao.deleteById(id);
     }
 
     private boolean isUserExist(User user) {
